Refactor pgwire server to better honour protocol
Commit refactors pgwire server so that bind and parse happen as explicit
requests to the xt node, rather than being purely pgwire server operations,
with only cmd-exec interacting with the node.

This change enables us to have more information about the query earlier
in the protocol, such as knowing the types and names of return columns
after the bind phase, rather than the execute phase. Previously we were
inferring this information from the SQL string itself, which has
limitations. As a result, SELECT * should now work over pgwire.

For details please see issue #3266

Current implementation differs from that outlined in the above issue in
that it chooses to do both parse and bind steps during bind, as this
enables us to reuse the same query-opts for both requests, avoiding the
need to check if the plan produced in parse is still valid when calling
bind.